Refactoring function calculateRollingAverage The rolling average is just the calculate of n elements given. The array bpiList stored all the values getting of the response and the array grows exponentially. The buffer got the elements of a stream of the array. this method to calculate the average consuming many resources.
The solution:
push the elments in the array, when have a lenght of five elements, calculate the rolling average and
shift the first element of the array.
when the next value is comming then push the value at the array.
in this manner the array ever will have 5 elements and the calculation will be faster.